remove debug component F0rm
