+++ docs/examples/.eslintrc
@@ -7,7 +7,6 @@
     "React": false,
     "ReactDOM": false,
 
-    "F0rm": false,
     "Form": false,
     "Grid": false,
     "Refer": false,
