+++ docs/src/Samples.js
@@ -2,7 +2,6 @@
 
 // 请按照字母顺序进行排序
 export default {
-  F0rmBasic:      require('fs').readFileSync(__dirname + '/../examples/F0rmBasic.js', 'utf8'),
   FormBasic:      require('fs').readFileSync(__dirname + '/../examples/FormBasic.js', 'utf8'),
   GridBasic:      require('fs').readFileSync(__dirname + '/../examples/GridBasic.js', 'utf8'),
   GridOperation:  require('fs').readFileSync(__dirname + '/../examples/GridOperation.js', 'utf8'),
